fixed app crashing when not logged in; need to implement anonymous logging

+++ app/src/main/java/com/yonasoft/jadedictionary/presentation/screens/lists/ListsViewModel.kt
@@ -7,16 +7,19 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.google.firebase.Firebase
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.auth
 import com.yonasoft.jadedictionary.data.datastore.StoreSearchHistory
 import com.yonasoft.jadedictionary.data.enums.SortOption
 import com.yonasoft.jadedictionary.data.models.Word
 import com.yonasoft.jadedictionary.data.models.WordList
+import com.yonasoft.jadedictionary.data.respositories.FirebaseAuthRepository
 import com.yonasoft.jadedictionary.data.respositories.WordListRepository
 import com.yonasoft.jadedictionary.data.respositories.WordRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
@@ -25,11 +28,16 @@ import javax.inject.Inject
 
 @HiltViewModel
 class ListsViewModel @Inject constructor(
-    private val wordRepository: WordRepository,
+
+    private val firebaseAuthRepository: FirebaseAuthRepository,
     private val wordListRepository: WordListRepository,
     private val storeSearchHistory: StoreSearchHistory,
 ) :
     ViewModel() {
+
+    private val _isLoggedIn = MutableStateFlow(false)
+    val isLoggedIn: StateFlow<Boolean> = _isLoggedIn.asStateFlow()
+
     val query = mutableStateOf("")
     val isActive = mutableStateOf(false)
     val isSyncing = mutableStateOf(false)
@@ -45,8 +53,18 @@ class ListsViewModel @Inject constructor(
 
 
     init {
-        getAllWordList()
         getHistory()
+        firebaseAuthRepository.getAuth().addAuthStateListener { auth ->
+            _isLoggedIn.value = auth.currentUser != null
+            if (_isLoggedIn.value) {
+                // User is logged in, perform actions that require authentication
+                getAllWordList()
+            } else {
+                // User is not logged in, clear sensitive data or adjust UI accordingly
+                _wordLists.value = emptyList()
+                _history.value = emptyList()
+            }
+        }
     }
 
     private fun getHistory() {
@@ -94,13 +112,14 @@ class ListsViewModel @Inject constructor(
     }
 
     private fun getAllWordList() {
-        viewModelScope.launch {
-            wordListRepository.getWordLists().collect{
-                _wordLists.value = it
+        if (_isLoggedIn.value) {
+            viewModelScope.launch {
+                wordListRepository.getWordLists().collect { lists ->
+                    _wordLists.value = lists
+                }
             }
         }
     }
-
     fun searchWordLists(searchQuery: String) {
         viewModelScope.launch {
             // Assuming getAllWordLists() can handle search queries or replace with appropriate search function
@@ -110,27 +129,33 @@ class ListsViewModel @Inject constructor(
     }
 
     fun addWordList(title: String = addTitle.value, description: String? = null) {
-        viewModelScope.launch {
-            val newWordList = WordList(
-                title = title,
-                description = description,
-                createdAt = Date(),
-                lastUpdatedAt = Date(),
-                userUid = Firebase.auth.currentUser?.uid ?: "",
-                wordIds = emptyList()
-            )
-
-            // Add to local database
-            wordListRepository.addOrUpdateWordList(newWordList)
+        if (_isLoggedIn.value) {
+            viewModelScope.launch {
+                val userUid = firebaseAuthRepository.getAuth().currentUser?.uid ?: return@launch
+                val newWordList = WordList(
+                    title = title,
+                    description = description,
+                    createdAt = Date(),
+                    lastUpdatedAt = Date(),
+                    userUid = userUid,
+                    wordIds = emptyList()
+                )
+                wordListRepository.addOrUpdateWordList(newWordList)
+            }
+        } else {
+            Log.d("ListsViewModel", "User must be logged in to add a word list.")
         }
     }
 
 
     fun deleteWordList(context: Context, wordList: WordList) {
-        viewModelScope.launch {
-            wordListRepository.deleteWordList(wordList.firebaseId)
-            Toast.makeText(context, "Word List: ${wordList.title} removed", Toast.LENGTH_SHORT).show()
+        if (_isLoggedIn.value) {
+            viewModelScope.launch {
+                wordListRepository.deleteWordList(wordList.firebaseId)
+                Toast.makeText(context, "Word List: ${wordList.title} removed", Toast.LENGTH_SHORT).show()
+            }
+        } else {
+            Log.d("ListsViewModel", "User must be logged in to delete a word list.")
         }
     }
-
 }
